requirements.txt
pydantic
fastapi==0.72.0
FastAPI-SQLAlchemy==0.2.1
uvicorn
gunicorn
python-multipart
python-jose[cryptography]
passlib[bcrypt]
phonenumbers
